Creada nueva excepcion personalizada para las autenticaciones Y limpiado codigo de autenticacion en VolunteerController y PartnerEntityController
