+++ backend/src/main/java/es/grupo8/backend/exceptions/AuthException.java
@@ -1,0 +1,17 @@
+package es.grupo8.backend.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+public class AuthException extends RuntimeException {
+
+    private final HttpStatus status;
+
+    public AuthException(HttpStatus status, String message) {
+        super(message);
+        this.status = status;
+    }
+
+    public HttpStatus getStatus() {
+        return status;
+    }
+}
